Devices/Audio/testcase/tstAudioMixBuffer.cpp: Make it work for hosts where we don't know the host page size when building, bugref:10391 [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@108633 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/testcase/tstAudioMixBuffer.cpp
@@ -168,7 +168,7 @@ static void tstBasics(RTTEST hTest)
     RTTESTI_CHECK_MSG((u32 = PDMAudioPropsMilliToBytes(&Cfg441StereoS16,          5)) == 884,       ("cb=%RU32\n", u32));
 
     /* DrvAudioHlpClearBuf: */
-    size_t const cbPageSize = RTSystemGetPageSize();
+    uint32_t const cbPageSize = RTSystemGetPageSize();
     uint8_t *pbPage;
     int rc = RTTestGuardedAlloc(hTest, cbPageSize, 0, false /*fHead*/, (void **)&pbPage);
     RTTESTI_CHECK_RC_OK_RETV(rc);
